+++ src/org/goda/chronic/repeaters/RepeaterMinute.java
@@ -55,10 +55,10 @@ public class RepeaterMinute extends RepeaterUnit {
   }
 
   @Override
-  public MutableInterval getOffset(MutableInterval span, int amount, Pointer.PointerType pointer) {
+  public MutableInterval getOffset(MutableInterval span, double amount, Pointer.PointerType pointer) {
     int direction = (pointer == Pointer.PointerType.FUTURE) ? 1 : -1;
     // WARN: Does not use Calendar
-    int seconds = direction * amount * RepeaterMinute.MINUTE_SECONDS;
+    int seconds = (int) (direction * amount * RepeaterMinute.MINUTE_SECONDS);
     return new MutableInterval( span.getStart().plusSeconds(seconds), span.getEnd().plusSeconds(seconds));
   }
