+++ src/org/goda/chronic/repeaters/RepeaterSeasonName.java
@@ -26,7 +26,7 @@ public class RepeaterSeasonName extends Repeater<Object> {
   }
 
   @Override
-  public MutableInterval getOffset(MutableInterval span, int amount, PointerType pointer) {
+  public MutableInterval getOffset(MutableInterval span, double amount, PointerType pointer) {
     throw new IllegalStateException("Not implemented.");
   }
